WB-3411 Prep for Java unit testing

+++ lib_builder/build.sh
@@ -24,7 +24,7 @@ if [ "$distro" -ne 0 ]; then
 else
     export DISTRO="CentOS 7"
     export BUILDENV="ncbi"
-    export JAVA_HOME=$(dirname $(dirname $(readlink -f $(which javac))))
+    export JAVA_HOME=$(dirname $(dirname $(readlink -f $(command -v javac))))
     export LD_LIBRARY_PATH="$LD_LIBRARY_PATH:/opt/ncbi/gcc/4.9.3/lib64/"
     export BLASTDB=/net/frosty/vol/blast/db/blast
     BLASTBYDATE="/panfs/pan1.be-md.ncbi.nlm.nih.gov/blastprojects/blast_build/c++/"
@@ -55,55 +55,66 @@ set -o errexit
 MAIN_JAR="../pipeline/target/sparkblast-1-jar-with-dependencies.jar"
 DEPENDS="$SPARK_HOME/jars/*:$MAIN_JAR:."
 
-pushd ../pipeline > /dev/null
-if [ "0" == "1" ]; then
+#TS=$(date +"%Y-%m-%d_%H%M%S")
+if [ "1" == "1" ]; then
+    pushd ../pipeline > /dev/null
     echo "Compiling Java"
-        #gsutil mb -p ncbi-sandbox-blast -c regional -l us-east4 gs://blast-builds
-        echo '{ "rule": [ { "action": {"type": "Delete"}, "condition": {"age": 7} } ] }' >lifecycle.json
-        echo '{ "description": "static_analysis_results", "owner" : "vartanianmh" }' > labels.json
-        #gsutil lifecycle set rule.json gs://blast-builds
-        #gsutil label set labels.json gs://blast-builds
-        TS=$(date +"%Y-%m-%d_%H%M%S")
+    #gsutil mb -p ncbi-sandbox-blast -c regional -l us-east4 gs://blast-builds
+    echo '{ "rule": [ { "action": {"type": "Delete"}, "condition": {"age": 7} } ] }' >lifecycle.json
+    echo '{ "description": "static_analysis_results", "owner" : "vartanianmh" }' > labels.json
+    #gsutil lifecycle set rule.json gs://blast-builds
+    #gsutil label set labels.json gs://blast-builds
 
     #../lib_builder/protoc -I../specs/ --java_out=. blast_request.proto
     #../lib_builder/protoc -I../specs/ --python_out=../tests/ blast_request.proto
-    #mvn compile
-
-    ./make_jar.sh
+    mvn compile
+    #    ./make_jar.sh
+    popd
 fi
 
-if [ "0" == "1" ]; then
+if [ "1" == "1" ]; then
+    pushd ../pipeline > /dev/null
+    set +errexit
     echo
     echo "Running Java linters/static analyzers"
+    echo "    Running checkstyle"
     mvn -q checkstyle:checkstyle > /dev/null 2>&1
-    CHECKSTYLE="gs://blast-builds/checkstyle_sun.$TS.html"
-#    gsutil cp target/site/checkstyle.html "$CHECKSTYLE"
-#    echo "  Output in $CHECKSTYLE"
+    #CHECKSTYLE="gs://blast-builds/checkstyle_sun.$TS.html"
+    #    gsutil cp target/site/checkstyle.html "$CHECKSTYLE"
+    cp target/site/checkstyle.html checkstyle.html
+    #    echo "  Output in $CHECKSTYLE"
 
     #mvn -q site > /dev/null 2>&1
+
+#    alias pmd="$HOME/pmd-bin-6.12.0/bin/run.sh pmd"
+#    type pmd
+echo "    Running PMD"
+#if type pmd; then
+"$HOME/pmd-bin-6.12.0/bin/run.sh" pmd -f textcolor -dir ./src/ -rulesets rulesets/java/quickstart.xml,category/java/codestyle.xml,category/java/bestpractices.xml > pmd.txt 2>/dev/null
+PMDLINES=$(wc -l < pmd.txt)
+echo "PMD had $PMDLINES complaints"
+#else
+#    echo "Skipping pmd"
+#fi
+set -o errexit
+
     #PMD="gs://blast-builds/pmd.$TS.html"
     #gsutil cp target/site/pmd.html "$PMD"
 
-    CHECKSTYLE="gs://blast-builds/checkstyle_google.$TS.html"
+#    CHECKSTYLE="gs://blast-builds/checkstyle_google.$TS.html"
+
 #    gsutil cp target/site/checkstyle.html "$CHECKSTYLE"
 #    echo "  Output in $CHECKSTYLE"
-fi
 popd > /dev/null
+fi
 
 #NOTE: javah deprecated in Java 9, removed in Java 10
 #JAVASRCDIR="../pipeline/src/main/java"
 #    $JAVASRCDIR/BLAST_REQUEST.java \
-    #    $JAVASRCDIR/BLAST_PARTITION.java \
-    #    $JAVASRCDIR/BLAST_HSP_LIST.java \
-    #    $JAVASRCDIR/BLAST_TB_LIST.java \
-    #    $JAVASRCDIR/BLAST_LIB.java \
-
-if [ "0" == "1" ]; then
-    javac -Xlint:all -Xlint:-path -Xlint:-serial -cp "$DEPENDS":. -d . -h . \
-        ./BLAST_TEST.java
-    javac -Xlint:all -Xlint:-path -Xlint:-serial -cp "$DEPENDS":. -d . -h . \
-        ./BLAST_BENCH.java
-fi
+#    $JAVASRCDIR/BLAST_PARTITION.java \
+#    $JAVASRCDIR/BLAST_HSP_LIST.java \
+#    $JAVASRCDIR/BLAST_TB_LIST.java \
+#    $JAVASRCDIR/BLAST_LIB.java \
 
 echo
 echo "Creating JNI header"
@@ -114,13 +125,13 @@ javap -p -s ../pipeline/target/classes/gov/nih/nlm/ncbi/blastjni/BLAST_LIB.class
 javap -p -s ../pipeline/target/classes/gov/nih/nlm/ncbi/blastjni/BLAST_HSP_LIST.class >> signatures
 javap -p -s ../pipeline/target/classes/gov/nih/nlm/ncbi/blastjni/BLAST_TB_LIST.class >> signatures
 
-echo "Creating JAR"
+#echo "Creating JAR"
 #jar cf $MAIN_JAR gov/nih/nlm/ncbi/blastjni/*class
-cp $MAIN_JAR .
+#cp $MAIN_JAR .
 #rm -rf gov
 
 if [ "$BUILDENV" = "ncbi" ]; then
-#    rm -f libblastjni.o ../pipeline/libblastjni.so
+    #    rm -f libblastjni.o ../pipeline/libblastjni.so
     # Note: Library order important
     #       Hidden dl_open for libdw
     # Eugene has:
@@ -132,206 +143,88 @@ if [ "$BUILDENV" = "ncbi" ]; then
     g++ --version
     GPPCOMMAND="
     g++ \
-    blastjni.cpp \
-    -std=gnu++11 \
-    -Wall -O -I . \
-    -Wextra -pedantic \
-    -Wlogical-op \
-    -Wshadow \
-    -Wformat=2 \
-    -Wformat-security \
-    -Woverloaded-virtual \
-    -Wcast-align \
-    -Wno-ctor-dtor-privacy \
-    -Wdisabled-optimization \
-    -Winit-self \
-    -Wmissing-declarations \
-    -Wmissing-include-dirs \
-    -Wredundant-decls \
-    -Wsign-promo \
-    -Wstrict-overflow=5 \
-    -Wswitch \
-    -Wno-unused \
-    -Wnon-virtual-dtor \
-    -Wreorder \
-    -Wdeprecated \
-    -Wno-float-equal \
-    -shared \
-    -fPIC \
-    $JAVA_INC \
-    -L./int/blast/libs \
-    -I $BLASTBYDATE/include \
-    -I $BLASTBYDATE/ReleaseMT/inc \
-    -L $BLASTBYDATE/ReleaseMT/lib \
-    -I/panfs/pan1.be-md.ncbi.nlm.nih.gov/blastprojects/blast_build/lmdb-0.9.21 \
-    -L/panfs/pan1.be-md.ncbi.nlm.nih.gov/blastprojects/blast_build/lmdb-0.9.21 \
-    -L . \
-    -L ext \
-    -fopenmp -lxblastformat -lalign_format -ltaxon1 -lblastdb_format \
-    -lgene_info -lxformat -lxcleanup -lgbseq -lmlacli \
-    -lmla -lmedlars -lpubmed -lvalid -ltaxon3 -lxalnmgr \
-    -lblastxml -lblastxml2 -lxcgi -lxhtml -lproteinkmer \
-    -lxblast -lxalgoblastdbindex -lcomposition_adjustment \
-    -lxalgodustmask -lxalgowinmask -lseqmasks_io -lseqdb \
-    -lblast_services -lxalnmgr -lxobjutil -lxobjread \
-    -lvariation -lcreaders -lsubmit -lxnetblastcli \
-    -lxnetblast -lblastdb -lscoremat -ltables -lxregexp \
-    -lncbi_xloader_genbank -lncbi_xreader_id1 \
-    -lncbi_xreader \
-    -lncbi_xreader_id2 \
-    -lxconnect -lid1 -lid2 -lxobjmgr \
-    -lgenome_collection -lseqedit -lseqsplit -lsubmit \
-    -lseqset -lseq -lseqcode -lsequtil -lpub -lmedline \
-    -lbiblio -lgeneral -lxser -lxutil -lxncbi -lxcompress \
-    -llmdb-static -lpthread -lz -lbz2 \
-    -L/netopt/ncbi_tools64/lzo-2.05/lib64 \
-    -llzo2 -ldl -lz -lnsl -lrt -ldl -lm -lpthread \
-    -o ./libblastjni.so"
-echo $GPPCOMMAND
-    BLAST_SERVER_GPP_COMMAND="
-    g++ \
-    blast_server.cpp \
-    -std=gnu++11 \
-    -march=sandybridge \
-    -Wall -O3 -I . \
-    -Wextra -pedantic \
-    -Wlogical-op \
-    -Wshadow \
-    -Wformat=2 \
-    -Wformat-security \
-    -Woverloaded-virtual \
-    -Wcast-align \
-    -Wno-ctor-dtor-privacy \
-    -Wdisabled-optimization \
-    -Winit-self \
-    -Wmissing-declarations \
-    -Wmissing-include-dirs \
-    -Wredundant-decls \
-    -Wsign-promo \
-    -Wstrict-overflow=5 \
-    -Wswitch \
-    -Wno-unused \
-    -Wnon-virtual-dtor \
-    -Wreorder \
-    -Wdeprecated \
-    -Wno-float-equal \
-    -fPIC \
-    -L./int/blast/libs \
-    -I $BLASTBYDATE/include \
-    $JAVA_INC \
-    -I $BLASTBYDATE/ReleaseMT/inc \
-    -L $BLASTBYDATE/ReleaseMT/lib \
-    -I/panfs/pan1.be-md.ncbi.nlm.nih.gov/blastprojects/blast_build/lmdb-0.9.21 \
-    -L/panfs/pan1.be-md.ncbi.nlm.nih.gov/blastprojects/blast_build/lmdb-0.9.21 \
-    -L . \
-    -L ext \
-    -L/netopt/ncbi_tools64/lzo-2.05/lib64 \
-    -fopenmp -lxblastformat -lalign_format -ltaxon1 -lblastdb_format \
-    -lgene_info -lxformat -lxcleanup -lgbseq -lmlacli \
-    -lmla -lmedlars -lpubmed -lvalid -ltaxon3 -lxalnmgr \
-    -lblastxml -lblastxml2 -lxcgi -lxhtml -lproteinkmer \
-    -lxblast -lxalgoblastdbindex -lcomposition_adjustment \
-    -lxalgodustmask -lxalgowinmask -lseqmasks_io -lseqdb \
-    -lblast_services -lxalnmgr -lxobjutil -lxobjread \
-    -lvariation -lcreaders -lsubmit -lxnetblastcli \
-    -lxnetblast -lblastdb -lscoremat -ltables -lxregexp \
-    -lncbi_xloader_genbank -lncbi_xreader_id1 \
-    -lncbi_xreader \
-    -lncbi_xreader_id2 \
-    -lxconnect -lid1 -lid2 -lxobjmgr \
-    -lgenome_collection -lseqedit -lseqsplit -lsubmit \
-    -lseqset -lseq -lseqcode -lsequtil -lpub -lmedline \
-    -lbiblio -lgeneral -lxser -lxutil -lxncbi -lxcompress \
-    -llmdb-static -lpthread -lz -lbz2 \
-    -L/netopt/ncbi_tools64/lzo-2.05/lib64 \
-    -llzo2 -ldl -lz -lnsl -lrt -ldl -lm -lpthread \
-    -o blast_server"
+        blastjni.cpp \
+        -std=gnu++11 \
+        -Wall -O -I . \
+        -Wextra -pedantic \
+        -Wlogical-op \
+        -Wshadow \
+        -Wformat=2 \
+        -Wformat-security \
+        -Woverloaded-virtual \
+        -Wcast-align \
+        -Wno-ctor-dtor-privacy \
+        -Wdisabled-optimization \
+        -Winit-self \
+        -Wmissing-declarations \
+        -Wmissing-include-dirs \
+        -Wredundant-decls \
+        -Wsign-promo \
+        -Wstrict-overflow=5 \
+        -Wswitch \
+        -Wno-unused \
+        -Wnon-virtual-dtor \
+        -Wreorder \
+        -Wdeprecated \
+        -Wno-float-equal \
+        -shared \
+        -fPIC \
+        $JAVA_INC \
+        -L./int/blast/libs \
+        -I $BLASTBYDATE/include \
+        -I $BLASTBYDATE/ReleaseMT/inc \
+        -L $BLASTBYDATE/ReleaseMT/lib \
+        -I/panfs/pan1.be-md.ncbi.nlm.nih.gov/blastprojects/blast_build/lmdb-0.9.21 \
+        -L/panfs/pan1.be-md.ncbi.nlm.nih.gov/blastprojects/blast_build/lmdb-0.9.21 \
+        -L . \
+        -L ext \
+        -fopenmp -lxblastformat -lalign_format -ltaxon1 -lblastdb_format \
+        -lgene_info -lxformat -lxcleanup -lgbseq -lmlacli \
+        -lmla -lmedlars -lpubmed -lvalid -ltaxon3 -lxalnmgr \
+        -lblastxml -lblastxml2 -lxcgi -lxhtml -lproteinkmer \
+        -lxblast -lxalgoblastdbindex -lcomposition_adjustment \
+        -lxalgodustmask -lxalgowinmask -lseqmasks_io -lseqdb \
+        -lblast_services -lxalnmgr -lxobjutil -lxobjread \
+        -lvariation -lcreaders -lsubmit -lxnetblastcli \
+        -lxnetblast -lblastdb -lscoremat -ltables -lxregexp \
+        -lncbi_xloader_genbank -lncbi_xreader_id1 \
+        -lncbi_xreader \
+        -lncbi_xreader_id2 \
+        -lxconnect -lid1 -lid2 -lxobjmgr \
+        -lgenome_collection -lseqedit -lseqsplit -lsubmit \
+        -lseqset -lseq -lseqcode -lsequtil -lpub -lmedline \
+        -lbiblio -lgeneral -lxser -lxutil -lxncbi -lxcompress \
+        -llmdb-static -lpthread -lz -lbz2 \
+        -L/netopt/ncbi_tools64/lzo-2.05/lib64 \
+        -llzo2 -ldl -lz -lnsl -lrt -ldl -lm -lpthread \
+        -o ./libblastjni.so"
+            echo "g++ CLI is $GPPCOMMAND"
 
-    BLAST_JSON_GPP_COMMAND="
-    g++ \
-    blast_json.cpp \
-    -std=gnu++11 \
-    -march=sandybridge \
-    -Wall -O3 -I . \
-    -Wextra -pedantic \
-    -Wlogical-op \
-    -Wshadow \
-    -Wformat=2 \
-    -Wformat-security \
-    -Woverloaded-virtual \
-    -Wcast-align \
-    -Wno-ctor-dtor-privacy \
-    -Wdisabled-optimization \
-    -Winit-self \
-    -Wmissing-declarations \
-    -Wmissing-include-dirs \
-    -Wredundant-decls \
-    -Wsign-promo \
-    -Wstrict-overflow=5 \
-    -Wswitch \
-    -Wno-unused \
-    -Wnon-virtual-dtor \
-    -Wreorder \
-    -Wdeprecated \
-    -Wno-float-equal \
-    -fPIC \
-    -L./int/blast/libs \
-    -I $BLASTBYDATE/include \
-    $JAVA_INC \
-    -I $BLASTBYDATE/ReleaseMT/inc \
-    -L $BLASTBYDATE/ReleaseMT/lib \
-    -I/panfs/pan1.be-md.ncbi.nlm.nih.gov/blastprojects/blast_build/lmdb-0.9.21 \
-    -L/panfs/pan1.be-md.ncbi.nlm.nih.gov/blastprojects/blast_build/lmdb-0.9.21 \
-    -L . \
-    -L ext \
-    -L/netopt/ncbi_tools64/lzo-2.05/lib64 \
-    -fopenmp -lxblastformat -lalign_format -ltaxon1 -lblastdb_format \
-    -lgene_info -lxformat -lxcleanup -lgbseq -lmlacli \
-    -lmla -lmedlars -lpubmed -lvalid -ltaxon3 -lxalnmgr \
-    -lblastxml -lblastxml2 -lxcgi -lxhtml -lproteinkmer \
-    -lxblast -lxalgoblastdbindex -lcomposition_adjustment \
-    -lxalgodustmask -lxalgowinmask -lseqmasks_io -lseqdb \
-    -lblast_services -lxalnmgr -lxobjutil -lxobjread \
-    -lvariation -lcreaders -lsubmit -lxnetblastcli \
-    -lxnetblast -lblastdb -lscoremat -ltables -lxregexp \
-    -lncbi_xloader_genbank -lncbi_xreader_id1 \
-    -lncbi_xreader \
-    -lncbi_xreader_id2 \
-    -lxconnect -lid1 -lid2 -lxobjmgr \
-    -lgenome_collection -lseqedit -lseqsplit -lsubmit \
-    -lseqset -lseq -lseqcode -lsequtil -lpub -lmedline \
-    -lbiblio -lgeneral -lxser -lxutil -lxncbi -lxcompress \
-    -llmdb-static -lpthread -lz -lbz2 \
-    -L/netopt/ncbi_tools64/lzo-2.05/lib64 \
-    -llzo2 -ldl -lz -lnsl -lrt -ldl -lm -lpthread \
-    -o blast_json"
+            if [ "1" == "1" ]; then
+                echo
+                echo "Running static analysis on C++ code"
+                cppcheck -q --enable=all --platform=unix64 blastjni.cpp
+                #        cppcheck -q --enable=all --platform=unix64 --std=c++11 blast_json.cpp
+                scan-build --use-analyzer /usr/local/llvm/7.0.0/bin/clang "$GPPCOMMAND"
+                #        scan-build --use-analyzer /usr/local/llvm/7.0.0/bin/clang "$BLAST_JSON_GPP_COMMAND"
+                #        scan-build --use-analyzer /usr/local/llvm/7.0.0/bin/clang "$BLAST_SERVER_GPP_COMMAND"
+                echo "Static analysis on C++ code complete"
+            fi
 
+            if [ "1" == "1" ]; then
+                echo
+                echo "Running clang-tidy checkers on C++ code"
+                /usr/local/llvm/7.0.0/bin/clang-tidy -checks='*,-cppcoreguidelines-pro-bounds-pointer-arithmetic,-cppcoreguidelines-pro-type-vararg,-hicpp-vararg,-fuchsia-default-arguments,-cppcoreguidelines-pro-bounds-array-to-pointer-decay,-hicpp-no-array-decay' \
+                    blastjni.cpp -- \
+                    -std=c++14 \
+                    -I. -I/usr/lib/jvm/java-1.8.0-openjdk-1.8.0.191.b12-0.el7_5.x86_64/include -I/usr/lib/jvm/java-1.8.0-openjdk-1.8.0.191.b12-0.el7_5.x86_64/include/linux -I/panfs/pan1.be-md.ncbi.nlm.nih.gov/blastprojects/blast_build/c++//include -I/panfs/pan1.be-md.ncbi.nlm.nih.gov/blastprojects/blast_build/c++//ReleaseMT/inc -I/panfs/pan1.be-md.ncbi.nlm.nih.gov/blastprojects/blast_build/lmdb-0.9.21 -I/usr/include/c++/4.8.2 -I/usr/include/c++/4.8.2/bits -I/usr/include/c++/4.8.2/x86_64-redhat-linux -I/usr/include/c++/4.8.2/backward
+                                    echo "Running clang-tidy checkers on C++ code"
+                                fi
 
-    if [ "1" == "1" ]; then
-        echo
-        echo "Running static analysis on C++ code"
-        cppcheck -q --enable=all --platform=unix64 blastjni.cpp
-#        cppcheck -q --enable=all --platform=unix64 --std=c++11 blast_json.cpp
-        scan-build --use-analyzer /usr/local/llvm/7.0.0/bin/clang "$GPPCOMMAND"
-#        scan-build --use-analyzer /usr/local/llvm/7.0.0/bin/clang "$BLAST_JSON_GPP_COMMAND"
-#        scan-build --use-analyzer /usr/local/llvm/7.0.0/bin/clang "$BLAST_SERVER_GPP_COMMAND"
-        echo "Static analysis on C++ code complete"
-    fi
-
-    if [ "1" == "1" ]; then
-        echo
-        echo "Running clang-tidy checkers on C++ code"
-        /usr/local/llvm/7.0.0/bin/clang-tidy -checks='*,-cppcoreguidelines-pro-bounds-pointer-arithmetic,-cppcoreguidelines-pro-type-vararg,-hicpp-vararg,-fuchsia-default-arguments,-cppcoreguidelines-pro-bounds-array-to-pointer-decay,-hicpp-no-array-decay' \
-            blastjni.cpp -- \
-            -std=c++14 \
-            -I. -I/usr/lib/jvm/java-1.8.0-openjdk-1.8.0.191.b12-0.el7_5.x86_64/include -I/usr/lib/jvm/java-1.8.0-openjdk-1.8.0.191.b12-0.el7_5.x86_64/include/linux -I/panfs/pan1.be-md.ncbi.nlm.nih.gov/blastprojects/blast_build/c++//include -I/panfs/pan1.be-md.ncbi.nlm.nih.gov/blastprojects/blast_build/c++//ReleaseMT/inc -I/panfs/pan1.be-md.ncbi.nlm.nih.gov/blastprojects/blast_build/lmdb-0.9.21 -I/usr/include/c++/4.8.2 -I/usr/include/c++/4.8.2/bits -I/usr/include/c++/4.8.2/x86_64-redhat-linux -I/usr/include/c++/4.8.2/backward
-        echo "Running clang-tidy checkers on C++ code"
-    fi
-
-    echo
-    echo "Compiling and linking blastjni.cpp"
-    $GPPCOMMAND
-    cp libblastjni.so ../pipeline
+                                echo
+                                echo "Compiling and linking blastjni.cpp"
+                                $GPPCOMMAND
+                                cp libblastjni.so ../pipeline
 
 #    $BLAST_JSON_GPP_COMMAND
 #    cp blast_json ../pipeline
@@ -382,31 +275,44 @@ if [[ $? -ne 1 ]]; then
 fi
 echo "  Testing for unresolved libraries OK"
 
-echo "  Testing JNI"
+pushd ../pipeline > /dev/null
+mvn package
+popd
+
 if [ "1" == "1" ]; then
-#-verbose:jni \
-    #-Djava.library.path="../pipeline" \
-    java \
-    -Xcheck:jni -Xdiag -Xfuture \
-    -cp $MAIN_JAR:. \
-    gov.nih.nlm.ncbi.blastjni.BLAST_TEST \
-    > output.$$ 2>&1
-    sort -u output.$$ | grep -e "^HSP: " -e "^TB: " > test.result
-    if ! cmp test.result test.expected
-    then
-        cat -tn output.$$
-        #rm -f output.$$
-        sdiff -w 70 test.result test.expected
-        echo "  Testing of JNI failed"
-        exit 1
-    fi
-fi
-rm -f output.$$
-set -o errexit
-echo "  Testing JNI OK"
-echo "Tests complete"
-line
-#fi
+    echo "Compiling JAVA test harnesses"
+    javac -Xlint:all -Xlint:-path -Xlint:-serial -cp "$DEPENDS":. -d . -h . \
+        ./BLAST_TEST.java
+            #    javac -Xlint:all -Xlint:-path -Xlint:-serial -cp "$DEPENDS":. -d . -h . \
+            #        ./BLAST_BENCH.java
+        fi
+
+
+        if [ "1" == "1" ]; then
+            echo "  Testing JNI"
+            #-verbose:jni \
+            #-Djava.library.path="../pipeline" \
+            java \
+                -Xcheck:jni -Xdiag -Xfuture \
+                -cp $MAIN_JAR:. \
+                gov.nih.nlm.ncbi.blastjni.BLAST_TEST \
+                > output.$$ 2>&1
+                            sort -u output.$$ | grep -e "^HSP: " -e "^TB: " > test.result
+                            if ! cmp test.result test.expected
+                            then
+                                cat -tn output.$$
+                                #rm -f output.$$
+                                sdiff -w 70 test.result test.expected
+                                echo "  Testing of JNI failed"
+                                exit 1
+                            fi
+                            echo "  Testing JNI OK"
+                        fi
+                        rm -f output.$$
+                        set -o errexit
+                        echo "Tests complete"
+                        line
+                        #fi
 
 #if [ "$BUILDENV" = "google" ]; then
 if [ "1" == "0" ]; then
@@ -432,7 +338,10 @@ if [ "1" == "0" ]; then
         --project ncbi-sandbox-blast \
         --config ../pipeline/cloudbuild.yaml \
         "--substitutions=_PATH=.,_TAG=dev-$BUILDTAG,_VERSION=$BUILDTAG" .
+
 fi
+
+popd > /dev/null
 
 echo "Build Complete"
 date
